requirements.txt
Django==2.0.6
djangorestframework==3.8.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.4
